+++ extension/src/experiments/columns/collect/util.ts
@@ -1,5 +1,5 @@
 import { Column, ColumnType } from '../../webview/contract'
-import { Value } from '../../../cli/dvc/reader'
+import { Value } from '../../../cli/dvc/contract'
 import { ConfigKey, getConfigValue } from '../../../vscode/config'
 
 export type ColumnAccumulator = Record<string, Column>
@@ -34,7 +34,7 @@ export const limitAncestorDepth = (
 } => {
   const [path, ...rest] = ancestors
   const collectedLimit = Number(
-    getConfigValue(ConfigKey.EXP_TABLE_HEAD_MAX_LAYERS, limit)
+    getConfigValue(ConfigKey.EXP_TABLE_HEAD_MAX_DEPTH, limit)
   )
 
   switch (collectedLimit) {