+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.13] - 2022-10-10
+
+### 🐛 Bug Fixes
+
+- Fix UX of extension using fallback global CLI when Python extension is active [#2544](https://github.com/iterative/vscode-dvc/pull/2544) by [@mattseddon](https://github.com/mattseddon)
+
+### 🔨 Maintenance
+
+- Add test for setting table header depth [#2525](https://github.com/iterative/vscode-dvc/pull/2525) by [@julieg18](https://github.com/julieg18)
+- Consolidate version checking into CLI discovery file [#2552](https://github.com/iterative/vscode-dvc/pull/2552) by [@mattseddon](https://github.com/mattseddon)
+
+## [0.4.12] - 2022-10-06
+
+### 🔨 Maintenance
+
+- Bump min tested version of DVC to 2.29.0 [#2536](https://github.com/iterative/vscode-dvc/pull/2536) by [@mattseddon](https://github.com/mattseddon)
+
+## [0.4.11] - 2022-10-05
+
+### 🚀 New Features and Enhancements
+
+- Allow experiment table column depth below 3 [#2482](https://github.com/iterative/vscode-dvc/pull/2482) by [@julieg18](https://github.com/julieg18)
+- Update table header context menu  [#2517](https://github.com/iterative/vscode-dvc/pull/2517) by [@julieg18](https://github.com/julieg18)
+
+### 🐛 Bug Fixes
+
+- Fix comparison table row path chevron being cut [#2533](https://github.com/iterative/vscode-dvc/pull/2533) by [@sroy3](https://github.com/sroy3)
+- Render flexible confusion matrices as expected [#2523](https://github.com/iterative/vscode-dvc/pull/2523) by [@mattseddon](https://github.com/mattseddon)
+- Fix size of flexible confusion matrix [#2531](https://github.com/iterative/vscode-dvc/pull/2531) by [@mattseddon](https://github.com/mattseddon)
+
+### 🔨 Maintenance
+
+- Move types from reader into new contract file [#2520](https://github.com/iterative/vscode-dvc/pull/2520) by [@mattseddon](https://github.com/mattseddon)
+- Switch e2e tests back to insiders build [#2526](https://github.com/iterative/vscode-dvc/pull/2526) by [@mattseddon](https://github.com/mattseddon)
+
 ## [0.4.10] - 2022-09-29
 
 ### 🐛 Bug Fixes